requirements.txt
discord.py
python-dotenv
youtube-search-python
yt-dlp
PyNaCl